+++ gmailconnector/__init__.py
@@ -12,9 +12,10 @@ from .read_email import ReadEmail  # noqa: F401
 from .send_email import SendEmail  # noqa: F401
 from .send_sms import SendSMS  # noqa: F401
 from .sms_deleter import DeleteSent  # noqa: F401
+from .validator.address import EmailAddress  # noqa: F401
 from .validator.validate_email import validate_email  # noqa: F401
 
-version = "0.7"
+version = "0.8"
 
 
 def load_env(filename: Union[str, os.PathLike] = ".env", scan: bool = False) -> NoReturn: